requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
aiosqlite==0.22.1
attrs==25.4.0
audioop-lts==0.2.2
discord.py==2.7.1
dotenv==0.9.9
frozenlist==1.8.0
humanfriendly==10.0
idna==3.11
multidict==6.7.1
propcache==0.4.1
pyreadline3==3.5.4
python-dotenv==1.2.2
yarl==1.23.0